+++ CHANGELOG.md
@@ -6,8 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [0.3.0]- 2019-06-06
+### Added
+- oauth2 authentication for endpoint calls
+- encrypted token saving to a file
+
+### Changed
+- client SDK configuration to be more flat
+
 ## [0.2.1]- 2019-05-23
-### Updated
+### Changed
 - symphony option-resolver version to be compatible with Mage 2.3+
 
 ## [0.2.0] - 2019-05-23
@@ -27,6 +35,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - createCategory event & direct logic
 - deleteCategory event & direct logic
 
-[Unreleased]: https://github.com/shopgate/cart-integration-sdk-php/compare/0.2.1...HEAD
+[Unreleased]: https://github.com/shopgate/cart-integration-sdk-php/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/shopgate/cart-integration-sdk-php/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/shopgate/cart-integration-sdk-php/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/shopgate/cart-integration-sdk-php/compare/0.1.0...0.2.0